+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- An index class for efficient in-memory access of sequence features.
 - Module for mRNA handling, with a function for selecting the primary mRNA from
   a gene or other feature.
 - New CLI command `tag pmrna`.